Move event-specific activities under the ui.eventShowcase, add intent to selectEvent button
