+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/eventShowcase/TicketActivity.java
@@ -1,13 +1,15 @@
-package ch.epfl.sweng.eventmanager;
+package ch.epfl.sweng.eventmanager.ui.eventShowcase;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class ScheduleActivity extends AppCompatActivity {
+import ch.epfl.sweng.eventmanager.R;
+
+public class TicketActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_schedule);
+        setContentView(R.layout.activity_ticket);
     }
 }
